requirements.txt
openai>=1.0.0
pandas
numpy
scikit-learn
xgboost
requests
beautifulsoup4
feedparser
ta
yfinance
flask
python-dotenv
twilio
gunicorn
pytest
